backend/ai_service.py
"""
AI Service - Multi-Provider AI Integration
Supports OpenAI, Google Gemini, and Anthropic Claude
"""

import asyncio
import json
from typing import Dict, List, Optional, Any
from datetime import datetime

# AI Provider imports
try:
    import openai
    OPENAI_AVAILABLE = True
except ImportError:
    OPENAI_AVAILABLE = False

try:
    import google.generativeai as genai
    GOOGLE_AVAILABLE = True
except ImportError:
    GOOGLE_AVAILABLE = False

try:
    import anthropic
    ANTHROPIC_AVAILABLE = True
except ImportError:
    ANTHROPIC_AVAILABLE = False

from config import settings

class AIService:
    """Multi-provider AI service for claim processing"""
    
    def __init__(self):
        self.providers = {
            "openai": OPENAI_AVAILABLE and bool(settings.OPENAI_API_KEY),
            "google": GOOGLE_AVAILABLE and bool(settings.GOOGLE_API_KEY),
            "anthropic": ANTHROPIC_AVAILABLE and bool(settings.ANTHROPIC_API_KEY),
            "demo": True  # Always available for demo mode
        }
        
        # Initialize providers
        if self.providers["openai"]:
            openai.api_key = settings.OPENAI_API_KEY
            
        if self.providers["google"]:
            genai.configure(api_key=settings.GOOGLE_API_KEY)
            
        if self.providers["anthropic"]:
            self.anthropic_client = anthropic.Anthropic(api_key=settings.ANTHROPIC_API_KEY)
    
    async def process_claim_document(self, document_text: str, document_type: str = "claim") -> Dict[str, Any]:
        """Process claim document with AI analysis"""
        
        if settings.DEMO_MODE or not any(self.providers.values()):
            return await self._demo_processing(document_text, document_type)
        
        # Try primary provider first
        provider = settings.DEFAULT_AI_PROVIDER
        
        if provider == "openai" and self.providers["openai"]:
            return await self._openai_processing(document_text, document_type)
        elif provider == "google" and self.providers["google"]:
            return await self._google_processing(document_text, document_type)
        elif provider == "anthropic" and self.providers["anthropic"]:
            return await self._anthropic_processing(document_text, document_type)
        else:
            # Fallback to demo mode
            return await self._demo_processing(document_text, document_type)
    
    async def _demo_processing(self, document_text: str, document_type: str) -> Dict[str, Any]:
        """Demo mode processing with realistic simulation"""
        
        # Simulate processing time
        await asyncio.sleep(2)
        
        # Generate realistic demo response
        demo_response = {
            "document_analysis": {
                "extracted_data": {
                    "claim_amount": "$2,450.00",
                    "policy_number": "POL-2024-001234",
                    "incident_date": "2024-01-15",
                    "claimant_name": "John Smith",
                    "incident_type": "Vehicle Collision",
                    "location": "123 Main St, Anytown, USA"
                },
                "confidence_score": 0.94,
                "processing_time": "2.1 seconds"
            },
            "fraud_detection": {
                "risk_score": 0.23,
                "risk_level": "Low",
                "fraud_indicators": [],
                "recommendation": "Approve for processing"
            },
            "ai_decision": {
                "recommendation": "Approve",
                "confidence": 0.89,
                "reasoning": "Standard claim with low fraud risk. All documentation appears valid.",
                "next_steps": ["Process payment", "Send confirmation"]
            },
            "processing_metadata": {
                "provider": "demo",
                "timestamp": datetime.now().isoformat(),
                "processing_time": 2.1,
                "model_version": "demo-v1.0"
            }
        }
        
        return demo_response
    
    async def _openai_processing(self, document_text: str, document_type: str) -> Dict[str, Any]:
        """OpenAI GPT-4 processing"""
        try:
            response = await openai.ChatCompletion.acreate(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": "You are an expert insurance claims processor. Analyze the claim document and provide structured analysis."},
                    {"role": "user", "content": f"Analyze this {document_type} document: {document_text[:1000]}..."}
                ],
                temperature=0.1
            )
            
            # Parse and structure the response
            return self._parse_ai_response(response.choices[0].message.content, "openai")
            
        except Exception as e:
            print(f"OpenAI processing error: {e}")
            return await self._demo_processing(document_text, document_type)
    
    async def _google_processing(self, document_text: str, document_type: str) -> Dict[str, Any]:
        """Google Gemini processing"""
        try:
            model = genai.GenerativeModel('gemini-pro')
            
            prompt = f"""
            Analyze this insurance claim document and provide structured analysis:
            
            Document Type: {document_type}
            Content: {document_text[:1000]}...
            
            Please provide:
            1. Extracted key information
            2. Fraud risk assessment
            3. Processing recommendation
            """
            
            response = await model.generate_content_async(prompt)
            
            return self._parse_ai_response(response.text, "google")
            
        except Exception as e:
            print(f"Google processing error: {e}")
            return await self._demo_processing(document_text, document_type)
    
    async def _anthropic_processing(self, document_text: str, document_type: str) -> Dict[str, Any]:
        """Anthropic Claude processing"""
        try:
            response = await self.anthropic_client.messages.create(
                model="claude-3-sonnet-20240229",
                max_tokens=1000,
                messages=[
                    {"role": "user", "content": f"Analyze this {document_type} document: {document_text[:1000]}..."}
                ]
            )
            
            return self._parse_ai_response(response.content[0].text, "anthropic")
            
        except Exception as e:
            print(f"Anthropic processing error: {e}")
            return await self._demo_processing(document_text, document_type)
    
    def _parse_ai_response(self, response_text: str, provider: str) -> Dict[str, Any]:
        """Parse AI response into structured format"""
        
        # This would parse the actual AI response
        # For now, return demo structure
        return {
            "document_analysis": {
                "extracted_data": {"status": "processed"},
                "confidence_score": 0.92,
                "processing_time": "1.8 seconds"
            },
            "fraud_detection": {
                "risk_score": 0.15,
                "risk_level": "Low",
                "fraud_indicators": [],
                "recommendation": "Approve for processing"
            },
            "ai_decision": {
                "recommendation": "Approve",
                "confidence": 0.91,
                "reasoning": f"AI analysis completed using {provider}",
                "next_steps": ["Process payment", "Send confirmation"]
            },
            "processing_metadata": {
                "provider": provider,
                "timestamp": datetime.now().isoformat(),
                "processing_time": 1.8,
                "model_version": f"{provider}-v1.0"
            }
        }
    
    def get_available_providers(self) -> Dict[str, bool]:
        """Get list of available AI providers"""
        return self.providers.copy()
    
    async def health_check(self) -> Dict[str, Any]:
        """Check health of all AI providers"""
        health_status = {
            "demo": True,
            "openai": False,
            "google": False,
            "anthropic": False
        }
        
        if self.providers["openai"]:
            try:
                # Test OpenAI connection
                health_status["openai"] = True
            except:
                pass
        
        if self.providers["google"]:
            try:
                # Test Google connection
                health_status["google"] = True
            except:
                pass
        
        if self.providers["anthropic"]:
            try:
                # Test Anthropic connection
                health_status["anthropic"] = True
            except:
                pass
        
        return health_status

# Global AI service instance
ai_service = AIService()
