backend/requirements.txt
# ClaimAI Pro Backend Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
supabase==2.7.4
pydantic==2.5.0

# AI Providers
openai==1.3.7
google-generativeai==0.3.2
anthropic==0.7.8

# Document Processing
pypdf2==3.0.1
pillow==10.1.0
python-multipart==0.0.6

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.2

# Development Tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0
